requirements.txt
python-telegram-bot==21.5
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0